requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.4
dataclasses-json==0.6.1
docopt==0.6.2
elastic-transport==8.4.1
elasticsearch==8.10.0
frozenlist==1.4.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.305
langsmith==0.0.41
markdownify==0.11.6
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
packaging==23.1
pipreqs==0.4.13
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.16
validators==0.22.0
weaviate-client==3.24.1
yarg==0.1.9
yarl==1.9.2
